requirements.txt
tk  # Python 自带
pyyaml
xmltodict
